+++ game/calcium-rendering-world3d/src/mesh.rs
@@ -31,13 +31,13 @@ impl Vertex {
 #[derive(Clone)]
 pub struct Mesh {
     pub vertices: Vec<Vertex>,
-    pub indices: Vec<u16>,
+    pub indices: Vec<u32>,
 }
 
 impl Mesh {
     /// Creates a mesh from vertices and indices. Performs no duplicate checking.
     pub fn new(
-        vertices: Vec<Vertex>, indices: Vec<u16>
+        vertices: Vec<Vertex>, indices: Vec<u32>
     ) -> Arc<Mesh> {
         Arc::new(Mesh {
             vertices,
@@ -48,7 +48,7 @@ impl Mesh {
 
 /// Converts a flat vertices vector to indexed vertices. Will eliminate duplicate vertices. Avoid
 /// using if you can directly provide vertices/indices without duplicate checking instead.
-pub fn flat_vertices_to_indexed(flat_vertices: &Vec<Vertex>) -> (Vec<Vertex>, Vec<u16>) {
+pub fn flat_vertices_to_indexed(flat_vertices: &Vec<Vertex>) -> (Vec<Vertex>, Vec<u32>) {
     let mut vertices = Vec::new();
     let mut indices = Vec::new();
     // There's no big significance to this capacity, other than leaving it at lowest is likely slow
@@ -66,8 +66,8 @@ pub fn flat_vertices_to_indexed(flat_vertices: &Vec<Vertex>) -> (Vec<Vertex>, Ve
 
 fn find_or_add_vertex(
     vertex: Vertex,
-    vertices: &mut Vec<Vertex>, indices: &mut Vec<u16>, lookup: &mut HashMap<u64, u16>,
-    i: &mut u16
+    vertices: &mut Vec<Vertex>, indices: &mut Vec<u32>, lookup: &mut HashMap<u64, u32>,
+    i: &mut u32
 ) {
     // Check if we found a matching vertex before. This is the reason we need a hash map of
     //  indices, a linear vector equality lookup would make this O(n^2). To generate the hash
